+++ conelight.cpp
@@ -16,7 +16,7 @@ Spectrum ConeLight::getIntensity(HitRecord &hit, QVector3D &direction, const Int
     shadowHit = scene.intersect(Ray(shadowHit.getIntersectingPoint(), -direction.normalized()), EPSILON, (shadowHit.getIntersectingPoint() - location).length());
     part *= (shadowHit.getIntersectingPoint() - location).length() * .001;
   }
-  double rayAngle = QVector3D::dotProduct(direction.normalized(), this->direction.normalized());
+  float rayAngle = QVector3D::dotProduct(direction.normalized(), this->direction.normalized());
   if(0 <= acos(rayAngle) && acos(rayAngle) <= openingAngle && !shadowHit.intersects())
   {
     return intensity * part / direction.lengthSquared();
@@ -27,7 +27,7 @@ Spectrum ConeLight::getIntensity(HitRecord &hit, QVector3D &direction, const Int
   }
 }
 
-Ray ConeLight::getRandomRay(const Sample &sample, double &pdf) const
+Ray ConeLight::getRandomRay(const Sample &sample, float &pdf) const
 {
   return Ray(location, sample.getCosineWeightedDirection(direction, pdf, openingAngle));
 }
