+++ intersectableinstance.h
@@ -15,7 +15,7 @@ public:
 
   }
 
-  HitRecord intersect(Ray ray, double from, double to) const;
+  HitRecord intersect(Ray ray, float from, float to) const;
   AxisAlignedBox * boundingBox() const;
 
 private:
