+++ phongmaterial.h
@@ -7,7 +7,7 @@
 class PhongMaterial : public DiffuseMaterial
 {
   public:
-    PhongMaterial(Spectrum color, Spectrum specularColor, double specularCoefficient) : color(color), specularColor(specularColor), specularCoefficient(specularCoefficient)
+    PhongMaterial(Spectrum color, Spectrum specularColor, float specularCoefficient) : color(color), specularColor(specularColor), specularCoefficient(specularCoefficient)
     {
 
     }
