+++ plane.cpp
@@ -4,11 +4,11 @@
 
 #include <limits>
 
-HitRecord Plane::intersect(Ray ray, double from, double to) const
+HitRecord Plane::intersect(Ray ray, float from, float to) const
 {
-  double d = -QVector3D::dotProduct(ray.getDirection(), planeVector.toVector3D());
-  double u = QVector4D::dotProduct(QVector4D(ray.getOrigin(), 1), planeVector);
-  double t = u / d;
+  float d = -QVector3D::dotProduct(ray.getDirection(), planeVector.toVector3D());
+  float u = QVector4D::dotProduct(QVector4D(ray.getOrigin(), 1), planeVector);
+  float t = u / d;
   if(from < t && t < to)
   {
     return HitRecord(t, ray, material, planeVector.toVector3D());
@@ -21,6 +21,6 @@ HitRecord Plane::intersect(Ray ray, double from, double to) const
 
 AxisAlignedBox* Plane::boundingBox() const
 {
-  double inf = std::numeric_limits< double >::infinity();
+  float inf = std::numeric_limits< float >::infinity();
   return new AxisAlignedBox(QVector3D(-inf, -inf, -inf), QVector3D(inf, inf, inf), DarkMatter::getInstance());
 }
