+++ pointlight.h
@@ -17,7 +17,7 @@ public:
 
 
   Spectrum getIntensity (const QVector3D & at, QVector3D & direction, const Intersectable & scene, const Sample &sample) const;
-  virtual Ray getRandomRay(const Sample& sample, double& pdf) const;
+  virtual Ray getRandomRay(const Sample& sample, float& pdf) const;
   QVector3D getDirection (QVector3D at) const;
   bool isOccluded(QVector3D location, const Intersectable & scene) const;
