+++ quad.cpp
@@ -1,6 +1,6 @@
 #include "quad.h"
 
-HitRecord Quad::intersect(Ray ray, double from, double to) const
+HitRecord Quad::intersect(Ray ray, float from, float to) const
 {
   HitRecord hit = t1.intersect(ray, from, to);
   if(hit.intersects())
