+++ quad.h
@@ -16,7 +16,7 @@ public:
     bBox.includePoint(p4);
   }
 
-  HitRecord intersect(Ray ray, double from, double to) const;
+  HitRecord intersect(Ray ray, float from, float to) const;
 
   AxisAlignedBox * boundingBox() const;
