+++ sampler.h
@@ -12,8 +12,8 @@ public:
   Sample(QPointF sample) : sample(sample) {}
 
   QPointF getSample() const;
-  QVector3D getCosineWeightedDirection(QVector3D w, double & pdf) const;
-  QVector3D getCosineWeightedDirection(QVector3D w, double &pdf, double openingAngle) const;
+  QVector3D getCosineWeightedDirection(QVector3D w, float & pdf) const;
+  QVector3D getCosineWeightedDirection(QVector3D w, float &pdf, float openingAngle) const;
   QVector3D getUniformSphereDirection() const;
 
 private:
@@ -26,7 +26,7 @@ public:
     virtual std::list<Sample> getSamples() = 0;
 
 protected:
-    static double getRandom();
+    static float getRandom();
 };
 
 inline QPointF Sample::getSample() const
