+++ transparentmaterial.cpp
@@ -8,7 +8,7 @@
 #include <cmath>
 #include <limits>
 
-TransparentMaterial::TransparentMaterial(double refractionCoefficient): MirrorMaterial(1), refractionCoefficient(refractionCoefficient)
+TransparentMaterial::TransparentMaterial(float refractionCoefficient): MirrorMaterial(1), refractionCoefficient(refractionCoefficient)
 {
 
 }
@@ -23,10 +23,10 @@ Spectrum TransparentMaterial::shade(HitRecord& hit, Light& light, const Intersec
   Spectrum mirrorComponent = MirrorMaterial::shade(hit, light, scene, depth + 1);
   QVector3D rayDirection = hit.getRay().getDirection().toVector3D().normalized();
   QVector3D surfaceNormal = hit.getSurfaceNormal().normalized();
-  double cosAngle = QVector3D::dotProduct(-surfaceNormal, rayDirection);
-  double angle = acos(cosAngle);
+  float cosAngle = QVector3D::dotProduct(-surfaceNormal, rayDirection);
+  float angle = acos(cosAngle);
   if(sin(angle) * refractionCoefficient > 1) return mirrorComponent;
-  double outAngle = asin(refractionCoefficient * sin(angle));
+  float outAngle = asin(refractionCoefficient * sin(angle));
   QVector3D refractedDirectionFront = refractionCoefficient * rayDirection + (refractionCoefficient * cosAngle - cos(outAngle)) * hit.getSurfaceNormal();
   QVector3D refractedDirectionBack = 1 / refractionCoefficient * rayDirection + (1 / refractionCoefficient * cosAngle - cos(outAngle)) * hit.getSurfaceNormal();
   refractedDirectionFront.normalize();
@@ -35,7 +35,7 @@ Spectrum TransparentMaterial::shade(HitRecord& hit, Light& light, const Intersec
   Ray newFrontRay = Ray(newOrigin, refractedDirectionFront);
   Ray newBackRay = Ray(newOrigin, -refractedDirectionBack);
   HitRecord newHitFront = scene.intersect(newFrontRay);
-  HitRecord newHitBack = scene.intersect(newBackRay, -EPSILON, -std::numeric_limits< double >::infinity());
+  HitRecord newHitBack = scene.intersect(newBackRay, -EPSILON, -std::numeric_limits< float >::infinity());
   HitRecord newHit;
   if(newHitFront.getRayParameter() < newHitBack.getRayParameter())
   {
@@ -45,8 +45,8 @@ Spectrum TransparentMaterial::shade(HitRecord& hit, Light& light, const Intersec
   {
     newHit = newHitBack;
   }
-  double f = pow(1 - refractionCoefficient, 2) / pow(1 + refractionCoefficient, 2);
-  double F = f + (1 - f) * pow(1 - QVector3D::dotProduct(rayDirection, -surfaceNormal), 5);
+  float f = pow(1 - refractionCoefficient, 2) / pow(1 + refractionCoefficient, 2);
+  float F = f + (1 - f) * pow(1 - QVector3D::dotProduct(rayDirection, -surfaceNormal), 5);
   return (1 - F) * newHit.getMaterial().shade(newHit, light, scene, depth + 1) + F * mirrorComponent;
 }
