+++ triangle.h
@@ -21,7 +21,7 @@ class Triangle : public Intersectable
     }
 
 
-    HitRecord intersect(Ray ray, double from, double to) const;
+    HitRecord intersect(Ray ray, float from, float to) const;
     AxisAlignedBox * boundingBox() const;
     
   private:
